% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosTrees.R
\docType{data}
\name{bosTrees}
\alias{bosTrees}
\title{Street trees in Boston}
\format{
A tibble with XXX rows and XXX variables:
\describe{
\item{FID}{unique identifier for each tree}
\item{Species}{a character denoting species}

}
}
\source{
{https://data.boston.gov/dataset/primary-street-trees-public}
}
\usage{
bosTrees
}
\description{
Includes data for street trees in Boston. Provided by Boston's GIS department.
}
\keyword{datasets}
