% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camTrees.R
\docType{data}
\name{camTrees}
\alias{camTrees}
\title{Street trees in Cambridge, MA}
\format{
A tibble with XXX rows and XXX variables:
\describe{
\item{XXX}{XXX}
\item{XXX}{XXX}

}
}
\source{
{https://www.cambridgema.gov/GIS/gisdatadictionary/Environmental/ENVIRONMENTAL_StreetTrees}
}
\usage{
camTrees
}
\description{
Includes data for street trees maintained by Cambridge, the Massachusetts Department of Conservation and Recreation, and private organizations.  The data are provided by Cambridge Public Works.
}
\keyword{datasets}
